+++ boat2023_nav2/params/costmap_params.yaml
@@ -1,14 +1,33 @@
-costmap_2d:
+global_costmap:
+  global_costmap:
+    ros__parameters:
+      update_frequency: 1.0
+      publish_frequency: 1.0
+      resolution: 0.05
+      robot_radius: 0.2
+      always_send_full_costmap: true
+      plugins: ["nav2_costmap_2d::InflationLayer", "nav2_gradient_costmap_plugin::GradientLayer"]
+
+local_costmap:
+  local_costmap:
+    ros__parameters:
+      update_frequency: 5.0
+      publish_frequency: 2.0
+      resolution: 0.05
+      robot_radius: 0.2
+      width: 4.0
+      height: 4.0
+      always_send_full_costmap: true
+      plugins: ["nav2_costmap_2d::InflationLayer", "nav2_gradient_costmap_plugin::GradientLayer"]
+
+gradient_layer:
   ros__parameters:
-    update_frequency: 5.0
-    publish_frequency: 2.0
-    global_frame: map
-    robot_base_frame: base_link
-    rolling_window: true
-    width: 10.0
-    height: 10.0
-    resolution: 0.05
-    plugins: ["gradient_layer"]
-    gradient_layer:
-      plugin: "nav2_gradient_costmap_plugin::GradientLayer"
-      # Add any other necessary parameters for the gradient layer here
+    plugin: "nav2_gradient_costmap_plugin::GradientLayer"
+    enabled: true
+    # Add any other configuration parameters specific to the plugin.
+
+inflation_layer:
+  ros__parameters:
+    plugin: "nav2_costmap_2d::InflationLayer"
+    cost_scaling_factor: 3.0
+    inflation_radius: 0.55
