+++ boat2023_nav2/params/ekf_params.yaml
@@ -1,0 +1,24 @@
+ekf_localization:
+  ros__parameters:
+    use_sim_time: True
+    frequency: 30.0
+    sensor_timeout: 0.1
+    two_d_mode: True
+    map_frame: 'map'
+    odom_frame: 'odom'
+    base_link_frame: 'ego_vehicle'
+    world_frame: 'odom'
+    
+    imu0: 'carla/ego_vehicle/imu'
+    imu0_config: [False, False, False,
+                  True,  True,  True,
+                  False, False, False,
+                  False, False, False,
+                  False, False, False]
+    imu0_nodelay: False
+    imu0_differential: False
+    imu0_relative: False
+    imu0_queue_size: 10
+    imu0_pose_rejection_threshold: 0.0
+    imu0_twist_rejection_threshold: 0.0
+    imu0_linear_acceleration_rejection_threshold: 0.0
