--- boat2023_nav2/params/nav2_params.yaml
@@ -1,33 +1,0 @@
-global_costmap:
-  global_costmap:
-    ros__parameters:
-      update_frequency: 1.0
-      publish_frequency: 1.0
-      resolution: 0.05
-      robot_radius: 0.2
-      always_send_full_costmap: true
-      plugins: ["nav2_costmap_2d::InflationLayer", "nav2_gradient_costmap_plugin::GradientLayer"]
-
-local_costmap:
-  local_costmap:
-    ros__parameters:
-      update_frequency: 5.0
-      publish_frequency: 2.0
-      resolution: 0.05
-      robot_radius: 0.2
-      width: 4.0
-      height: 4.0
-      always_send_full_costmap: true
-      plugins: ["nav2_costmap_2d::InflationLayer", "nav2_gradient_costmap_plugin::GradientLayer"]
-
-gradient_layer:
-  ros__parameters:
-    plugin: "nav2_gradient_costmap_plugin::GradientLayer"
-    enabled: true
-    # Add any other configuration parameters specific to the plugin.
-
-inflation_layer:
-  ros__parameters:
-    plugin: "nav2_costmap_2d::InflationLayer"
-    cost_scaling_factor: 3.0
-    inflation_radius: 0.55
